wip - working on flushing out the Rte sub-system

+++ src/Cpl/Rte/PointCommon.h
@@ -27,7 +27,7 @@ class Base: public Rte::Point::Api
 {
 protected:
     /// Internal sequence number to track changes to the Point
-    uint32_t m_seqnum_;
+    uint16_t m_seqnum_;
 
     /// My membership changed flag/state (For a Point Container
     bool     m_changed_;
